Unit test run with C as well as C++ compiler to check for compilation.
