+++ unit/test-vss.c
@@ -30,9 +30,13 @@
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
+#if defined(__cplusplus)
 extern "C" {
 #include <cmocka.h>
 }
+#else
+#include <cmocka.h>
+#endif
 #include <arpa/inet.h>
 #include <errno.h>
 #include <string.h>
@@ -108,7 +112,7 @@ static void vss_interop_path(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -135,7 +139,7 @@ static void vss_data_uint8(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -163,7 +167,7 @@ static void vss_data_int8(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -191,7 +195,7 @@ static void vss_data_uint16(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -219,7 +223,7 @@ static void vss_data_int16(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -247,7 +251,7 @@ static void vss_data_uint32(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -275,7 +279,7 @@ static void vss_data_int32(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -302,7 +306,7 @@ static void vss_data_uint64(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -331,7 +335,7 @@ static void vss_data_int64(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -360,7 +364,7 @@ static void vss_data_bool(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -388,7 +392,7 @@ static void vss_data_float(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -416,7 +420,7 @@ static void vss_data_double(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -444,17 +448,17 @@ static void vss_data_string(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     char str_value[] = "10m/s2";
-    VssDataString_t vss_data_string {};
+    VssDataString_t vss_data_string = {0};
     vss_data_string.data = str_value;
     vss_data_string.data_length = strlen(str_value);
-    VssData_t data = {};
+    VssData_t data = {0};
     data.data_string = &vss_data_string;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_STRING);
@@ -484,17 +488,17 @@ static void vss_data_uint8_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     uint8_t uint8_arr_value[] = {5, 4, 3, 2, 1};
-    VssDataUint8Array_t vss_data_uint8_arr {};
+    VssDataUint8Array_t vss_data_uint8_arr = {0};
     vss_data_uint8_arr.data = uint8_arr_value;
     vss_data_uint8_arr.data_length = 5;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_uint8_array = &vss_data_uint8_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_UINT8_ARRAY);
@@ -524,17 +528,17 @@ static void vss_data_int8_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     int8_t int8_arr_value[] = {-5, -4, -3, -2, -1};
-    VssDataInt8Array_t vss_data_int8_arr {};
+    VssDataInt8Array_t vss_data_int8_arr = {0};
     vss_data_int8_arr.data = int8_arr_value;
     vss_data_int8_arr.data_length = 5;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_int8_array = &vss_data_int8_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_INT8_ARRAY);
@@ -543,8 +547,8 @@ static void vss_data_int8_array(void **state) {
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 15, len_in_mem, 2);
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, int8_arr_value, 5);
 
-    VssData_t vss_data_get;
-    VssDataInt8Array_t vss_data_int8_arr_get {};
+    VssData_t vss_data_get = {0};
+    VssDataInt8Array_t vss_data_int8_arr_get = {0};
     vss_data_int8_arr_get.data = 0;
     vss_data_get.data_int8_array = &vss_data_int8_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_INT8_ARRAY);
@@ -563,17 +567,17 @@ static void vss_data_uint16_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     uint16_t uint16_arr_value[] = {0x100, 0x101, 0x102, 0x103, 0x104};
-    VssDataUint16Array_t vss_data_uint16_arr {};
+    VssDataUint16Array_t vss_data_uint16_arr = {0};
     vss_data_uint16_arr.data = uint16_arr_value;
     vss_data_uint16_arr.data_length = 10;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_uint16_array = &vss_data_uint16_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_UINT16_ARRAY);
@@ -584,9 +588,8 @@ static void vss_data_uint16_array(void **state) {
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 10);
 
     VssData_t vss_data_get;
-    VssDataUint16Array_t vss_data_uint16_arr_get = {
-        .data = 0
-    };
+    VssDataUint16Array_t vss_data_uint16_arr_get = {0};
+    vss_data_uint16_arr_get.data = 0;
     vss_data_get.data_uint16_array = &vss_data_uint16_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_UINT16_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -604,17 +607,17 @@ static void vss_data_int16_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     int16_t int16_arr_value[] = {-0x100, -0x101, -0x102, -0x103, -0x104};
-    VssDataInt16Array_t vss_data_int16_arr {};
+    VssDataInt16Array_t vss_data_int16_arr = {0};
     vss_data_int16_arr.data = int16_arr_value;
     vss_data_int16_arr.data_length = 10;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_int16_array = &vss_data_int16_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_INT16_ARRAY);
@@ -624,10 +627,9 @@ static void vss_data_int16_array(void **state) {
     uint8_t arr_in_mem[] = {0xFF, 0, 0xFE, 0xFF, 0xfe, 0xfe, 0xfe, 0xfd, 0xfe, 0xfc};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 10);
 
-    VssData_t vss_data_get;
-    VssDataInt16Array_t vss_data_int16_arr_get = {
-        .data = 0
-    };
+    VssData_t vss_data_get = {0};
+    VssDataInt16Array_t vss_data_int16_arr_get = {0};
+    vss_data_int16_arr_get.data = 0;
     vss_data_get.data_int16_array = &vss_data_int16_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_INT16_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -645,17 +647,17 @@ static void vss_data_uint32_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     uint32_t uint32_arr_value[] = {0x10000, 0x10100, 0x10200, 0x10300, 0x10400};
-    VssDataUint32Array_t vss_data_uint32_arr {};
+    VssDataUint32Array_t vss_data_uint32_arr = {0};
     vss_data_uint32_arr.data = uint32_arr_value;
     vss_data_uint32_arr.data_length = 20;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_uint32_array = &vss_data_uint32_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_UINT32_ARRAY);
@@ -665,10 +667,9 @@ static void vss_data_uint32_array(void **state) {
     uint8_t arr_in_mem[] = {0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 2, 0, 0, 1, 3, 0, 0, 1, 4, 0};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 20);
 
-    VssData_t vss_data_get;
-    VssDataUint32Array_t vss_data_uint32_arr_get = {
-        .data = 0
-    };
+    VssData_t vss_data_get = {0};
+    VssDataUint32Array_t vss_data_uint32_arr_get = {0};
+    vss_data_uint32_arr_get.data = 0;
     vss_data_get.data_uint32_array = &vss_data_uint32_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_UINT32_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -686,17 +687,17 @@ static void vss_data_int32_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     int32_t int32_arr_value[] = {-0x10000, -0x10100, -0x10200, -0x10300, -0x10400};
-    VssDataInt32Array_t vss_data_int32_arr {};
+    VssDataInt32Array_t vss_data_int32_arr = {0};
     vss_data_int32_arr.data = int32_arr_value;
     vss_data_int32_arr.data_length = 20;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_int32_array = &vss_data_int32_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_INT32_ARRAY);
@@ -710,10 +711,9 @@ static void vss_data_int32_array(void **state) {
                             0xff, 0xfe, 0xfc, 0};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 20);
 
-    VssData_t vss_data_get;
-    VssDataInt32Array_t vss_data_int32_arr_get = {
-        .data = 0
-    };
+    VssData_t vss_data_get = {0};
+    VssDataInt32Array_t vss_data_int32_arr_get = {0};
+    vss_data_int32_arr_get.data = 0;
     vss_data_get.data_int32_array = &vss_data_int32_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_INT32_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -731,18 +731,18 @@ static void vss_data_uint64_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     uint64_t uint64_arr_value[] = {0x10000, 0x10100, 0x10200, 0x10300, 0x10400};
-    VssDataUint64Array_t vss_data_uint64_arr {};
+    VssDataUint64Array_t vss_data_uint64_arr = {0};
     vss_data_uint64_arr.data = uint64_arr_value;
     vss_data_uint64_arr.data_length = 40;
 
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_uint64_array = &vss_data_uint64_arr;
     Avtp_Vss_SetDatatype(vss_pdu, VSS_UINT64_ARRAY);
     Avtp_Vss_SetVssData(vss_pdu, &data);
@@ -755,8 +755,8 @@ static void vss_data_uint64_array(void **state) {
                              0, 0, 0, 0, 0, 1, 4, 0};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 40);
 
-    VssData_t vss_data_get;
-    VssDataUint64Array_t vss_data_uint64_arr_get {};
+    VssData_t vss_data_get = {0};
+    VssDataUint64Array_t vss_data_uint64_arr_get = {0};
     vss_data_get.data_uint64_array = &vss_data_uint64_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_UINT64_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -774,17 +774,17 @@ static void vss_data_int64_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     int64_t int64_arr_value[] = {-0x10000, -0x10100, -0x10200, -0x10300, -0x10400};
-    VssDataInt64Array_t vss_data_int64_arr {};
+    VssDataInt64Array_t vss_data_int64_arr = {0};
     vss_data_int64_arr.data = int64_arr_value;
     vss_data_int64_arr.data_length = 40;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_int64_array = &vss_data_int64_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_INT64_ARRAY);
@@ -798,8 +798,8 @@ static void vss_data_int64_array(void **state) {
                              0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xfc, 0,};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 40);
 
-    VssData_t vss_data_get;
-    VssDataInt64Array_t vss_data_int64_arr_get {};
+    VssData_t vss_data_get = {0};
+    VssDataInt64Array_t vss_data_int64_arr_get = {0};
     vss_data_int64_arr_get.data = 0;
 
     vss_data_get.data_int64_array = &vss_data_int64_arr_get;
@@ -819,18 +819,18 @@ static void vss_data_bool_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     uint8_t bool_arr_value[] = {1, 0, 0, 0, 1};
-    VssDataBoolArray_t vss_data_bool_arr {};
+    VssDataBoolArray_t vss_data_bool_arr = {0};
     vss_data_bool_arr.data = bool_arr_value;
     vss_data_bool_arr.data_length = 5;
 
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_bool_array = &vss_data_bool_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_BOOL_ARRAY);
@@ -839,8 +839,8 @@ static void vss_data_bool_array(void **state) {
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 15, len_in_mem, 2);
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, bool_arr_value, 5);
 
-    VssData_t vss_data_get;
-    VssDataBoolArray_t vss_data_bool_arr_get {};
+    VssData_t vss_data_get = {0};
+    VssDataBoolArray_t vss_data_bool_arr_get = {0};
     vss_data_bool_arr_get.data = 0;
     vss_data_get.data_bool_array = &vss_data_bool_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_BOOL_ARRAY);
@@ -859,17 +859,17 @@ static void vss_data_float_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     float float_arr_value[] = {1.2, -1.2, 1.3, -1.3, 1.5};
-    VssDataFloatArray_t vss_data_float_arr {};
+    VssDataFloatArray_t vss_data_float_arr = {0};
     vss_data_float_arr.data = float_arr_value;
     vss_data_float_arr.data_length = 20;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_float_array = &vss_data_float_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_FLOAT_ARRAY);
@@ -883,8 +883,8 @@ static void vss_data_float_array(void **state) {
                              0x3f, 0xc0, 0, 0, 0};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 20);
 
-    VssData_t vss_data_get;
-    VssDataFloatArray_t vss_data_float_arr_get {};
+    VssData_t vss_data_get = {0};
+    VssDataFloatArray_t vss_data_float_arr_get = {0};
     vss_data_float_arr_get.data = 0;
 
     vss_data_get.data_float_array = &vss_data_float_arr_get;
@@ -904,17 +904,17 @@ static void vss_data_double_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
     Avtp_Vss_SetVssPath(vss_pdu, &path_id);
 
     double double_arr_value[] = {1.2, -1.2, 1.3, -1.3, 1.5};
-    VssDataDoubleArray_t vss_data_double_arr {};
+    VssDataDoubleArray_t vss_data_double_arr = {0};
     vss_data_double_arr.data = double_arr_value;
     vss_data_double_arr.data_length = 40;
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_double_array = &vss_data_double_arr;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_DOUBLE_ARRAY);
@@ -928,10 +928,8 @@ static void vss_data_double_array(void **state) {
                              0x3f, 0xf8, 0, 0, 0, 0, 0, 0};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 40);
 
-    VssData_t vss_data_get;
-    VssDataDoubleArray_t vss_data_double_arr_get = {
-        .data = 0
-    };
+    VssData_t vss_data_get = {0};
+    VssDataDoubleArray_t vss_data_double_arr_get = {0};
     vss_data_get.data_double_array = &vss_data_double_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_DOUBLE_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
@@ -949,7 +947,7 @@ static void vss_data_string_array(void **state) {
     char path[] = "Vehicle.Speed";
     uint32_t path_length = strlen(path);
 
-    VssPath_t path_id {};
+    VssPath_t path_id = {0};
     path_id.vss_interop_path.path = path;
     path_id.vss_interop_path.path_length = path_length;
     Avtp_Vss_SetAddrMode(vss_pdu, VSS_INTEROP_MODE);
@@ -958,23 +956,23 @@ static void vss_data_string_array(void **state) {
     char str1[] = "Hello";
     char str2[] = "World";
     char str3[] = "Tschuss";
-    VssDataString_t vss_str1 {};
+    VssDataString_t vss_str1 = {0};
     vss_str1.data = str1;
     vss_str1.data_length = strlen(str1);
-    VssDataString_t vss_str2 {};
+    VssDataString_t vss_str2 = {0};
     vss_str2.data = str2;
     vss_str2.data_length = strlen(str2);
-    VssDataString_t vss_str3 {};
+    VssDataString_t vss_str3 = {0};
     vss_str3.data = str3;
     vss_str3.data_length = strlen(str3);
 
     VssDataString_t* arr[] = {&vss_str1, &vss_str2, &vss_str3};
     uint8_t buffer[strlen(str1) + strlen(str2) + strlen(str3)+2];
-    VssDataStringArray_t vss_str_array {};
+    VssDataStringArray_t vss_str_array = {0};
     vss_str_array.data = buffer;
     Avtp_Vss_SerializeStringArray(&vss_str_array, arr, 3);
 
-    VssData_t data {};
+    VssData_t data = {0};
     data.data_string_array = &vss_str_array;
 
     Avtp_Vss_SetDatatype(vss_pdu, VSS_STRING_ARRAY);
@@ -986,10 +984,8 @@ static void vss_data_string_array(void **state) {
                              0x00, 0x07, 'T', 's', 'c', 'h', 'u', 's', 's'};
     assert_memory_equal(pdu + AVTP_VSS_FIXED_HEADER_LEN + 17, arr_in_mem, 23);
 
-    VssData_t vss_data_get;
-    VssDataStringArray_t vss_data_string_arr_get = {
-        .data = 0
-    };
+    VssData_t vss_data_get = {0};
+    VssDataStringArray_t vss_data_string_arr_get = {0};
     vss_data_get.data_string_array = &vss_data_string_arr_get;
     assert_int_equal(Avtp_Vss_GetDatatype(vss_pdu), VSS_STRING_ARRAY);
     Avtp_Vss_GetVssData(vss_pdu, &vss_data_get);
